api-gateway/requirements.txt
flask==1.1.2
jinja2==2.11.3
werkzeug==1.0.1
python-dotenv==0.15.0
requests==2.25.1
markupsafe==1.1.1
itsdangerous==1.1.0
flask_cors